Bugfix: Disable expire in flex output when not used
Origin: https://github.com/openstreetmap/osm2pgsql/issues/1436
Bug: https://github.com/openstreetmap/osm2pgsql/issues/1436

If we don't have expire enabled we never need to create an expire list.
But as a side effect of the two-stage processing we created such a list
which costs a lot of time and RAM. This commit disables this which
speeds up processing considerably.

Gbp-Pq: Name 0001-Bugfix-Disable-expire-in-flex-output-when-not-used.patch